pyproject.toml
[project]
name = "ai-rag-devops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.2",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "markdown>=3.10.2",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "unstructured>=0.21.5",
]
